+++ src/main/java/roomescape/controller/ReservationController.java
@@ -1,8 +1,11 @@
-package roomescape;
+package roomescape.controller;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
+import roomescape.dto.ReservationRequest;
+import roomescape.service.ReservationService;
+import roomescape.domain.Reservation;
 
 import java.util.List;
