+++ src/main/java/roomescape/service/ReservationService.java
@@ -1,6 +1,11 @@
-package roomescape;
+package roomescape.service;
 
 import org.springframework.stereotype.Service;
+import roomescape.dto.ReservationRequest;
+import roomescape.dao.ReservationDao;
+import roomescape.dao.ReservationTimeDao;
+import roomescape.domain.Reservation;
+import roomescape.domain.ReservationTime;
 
 import java.util.List;
