+++ src/test/java/roomescape/dao/ReservationDaoTest.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.dao;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -7,6 +7,8 @@ import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.datasource.DriverManagerDataSource;
 import org.springframework.jdbc.datasource.init.ResourceDatabasePopulator;
+import roomescape.domain.Reservation;
+import roomescape.domain.ReservationTime;
 
 import java.util.List;
